adcpi-knot-agent-main/requirements.txt
git+https://github.com/abelectronicsuk/ABElectronics_Python_Libraries.git
smbus2
paho-mqtt